refactor: move Backend to exported file
Getting confused everytime I need to look it up.
